[ci][microcheck/12] move get_new_tests and get_human_specified_tests to Test class
