+++ release/ray_release/tests/test_test.py
@@ -405,5 +405,16 @@ def test_get_changed_tests(
     assert Test.get_changed_tests("") == {"//t1", "//t2"}
 
 
+@mock.patch.dict(
+    os.environ,
+    {"BUILDKITE_PULL_REQUEST_BASE_BRANCH": "base", "BUILDKITE_COMMIT": "commit"},
+)
+@mock.patch("subprocess.check_call")
+@mock.patch("subprocess.check_output")
+def test_get_human_specified_tests(mock_check_output, mock_check_call) -> None:
+    mock_check_output.return_value = b"hi\n@microcheck //test01 //test02\nthere"
+    assert Test.get_human_specified_tests("") == {"//test01", "//test02"}
+
+
 if __name__ == "__main__":
     sys.exit(pytest.main(["-v", __file__]))
